pyproject.toml
[tool.poetry]
name = "llamafunctioncalling"
version = "0.1.0"
description = ""
authors = ["satyamdas03 <116027621+satyamdas03@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
psutil = "^6.1.0"
pydantic = "^2.10.1"
python-dotenv = "^1.0.1"
instructor = "^1.6.4"
openai = "^1.55.0"
speechrecognition = "^3.11.0"
pyttsx3 = "^2.98"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
